requirements.txt
ttkbootstrap
pillow